Moved random_projections under decompositions
