+++ sklearn/decomposition/tests/test_random_projection.py
@@ -3,9 +3,9 @@ import scipy.sparse as sp
 import numpy as np
 
 from sklearn.metrics import euclidean_distances
-from sklearn.random_projection import SparseRandomProjection
-from sklearn.random_projection import johnson_lindenstrauss_min_dim
-from sklearn.random_projection import random_dot
+from sklearn.decomposition.random_projection import SparseRandomProjection
+from sklearn.decomposition.random_projection import johnson_lindenstrauss_min_dim
+from sklearn.decomposition.random_projection import random_dot
 
 from sklearn.utils.testing import assert_raise_message
 from numpy.testing import assert_array_equal
